Noise A deep-dive rerun: direction exonerated, significance established, dissociation found
Rerun per preregistered spec (19 imgs x 3 conditions x 50 steps, seed 42):
- Historical values reproduced (orig 22.46 / cut_a 24.66 / noise_a 24.84)
- Paired t (noise vs orig): t=4.97, p=0.0001, d=0.66 — significantly better
- TOST (noise vs cut, 0.5 dB): p=0.0029 — equivalent
- Peak-layer drift: cut_a -27.7%, noise_a +6.4% — the pre-registered
  "double drop" condition was mis-specified; drift L2 measures trajectory
  discrepancy, not conflict. The rise is the capacity-preserved/topology-
  destroyed signature: conflict and discrepancy dissociate under noise.
- Fix NoiseIntervention dtype bug (fp32 mean/std cast back to t.dtype)
- Unify D_total -> D_s across docs and figure (definition-led naming)

+++ scripts/phase7_skip_noise_intervention.py
@@ -78,7 +78,7 @@ class NoiseIntervention:
                     # its per-tensor mean and std, preserving statistical
                     # properties but destroying spatial/feature structure
                     noisy = tuple(
-                        torch.randn_like(t) * t.std() + t.mean()
+                        (torch.randn_like(t) * t.std() + t.mean()).to(t.dtype)
                         for t in res_hidden_states_tuple
                     )
                     return orig_fn(hidden_states, noisy, *args, **kwargs)
